Moved the service functions out into a seperate file for the time being
